SOCAT/requirements.txt
astropy[recommended]
pyserial
protobuf
